renamed `SetupUserKeys.vue` → `InitialSetup.vue`
